requirements.txt
bcrypt==4.1.3
greenlet==3.0.3
grpc-interceptor==0.15.4
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.27.0
PyJWT==2.8.0
python-dotenv==1.0.1
setuptools==70.0.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
